Refactor summarizer to allow testing
