+++ src/keylogger.ts
@@ -8,8 +8,7 @@ import {
   IGlobalKeyEvent,
 } from "node-global-key-listener";
 import { appendFile, currentKeyLogFile } from "./electron/paths";
-import { loadPreferences } from "./main"; // Import the loadPreferences function
-import { Preferences } from "./preferences";
+import { loadPreferences, Preferences } from "./preferences";
 import log from "./logging";
 
 const BINARY_NAME = "MacKeyServer";
