+++ src/preferences.ts
@@ -1,3 +1,8 @@
+import fs from "node:fs/promises";
+import path from "path";
+
+import { app } from "electron";
+
 export interface ScreenshotPreferences {
   screenshotActive: boolean;
   screenshotPeriod: number;
@@ -44,3 +49,24 @@ export const DEFAULT_PREFERENCES: Preferences = {
     "Please analyze this computer activity log and summarize the major projects and tasks worked on:",
   summaryModel: "anthropic/claude-3.5-sonnet", // TODO select this from the available models
 };
+
+const userDataPath = app.getPath("userData");
+const preferencesPath = path.join(userDataPath, "preferences.json");
+
+export async function savePreferences(
+  prefs: Partial<Preferences>,
+): Promise<Preferences> {
+  const currentPrefs = await loadPreferences();
+  const newPrefs = { ...currentPrefs, ...prefs };
+  await fs.writeFile(preferencesPath, JSON.stringify(newPrefs, null, 2));
+  return newPrefs;
+}
+
+export async function loadPreferences(): Promise<Preferences> {
+  try {
+    const data = await fs.readFile(preferencesPath, "utf-8");
+    return { ...DEFAULT_PREFERENCES, ...JSON.parse(data) };
+  } catch {
+    return DEFAULT_PREFERENCES;
+  }
+}
